Creando clases Furniture y FurnitureRental en front/src/app/core/models
