+++ front/src/app/core/models/furniture-rental.spec.ts
@@ -1,0 +1,7 @@
+import { FurnitureRental } from './furniture-rental';
+
+describe('FurnitureRental', () => {
+  it('should create an instance', () => {
+    expect(new FurnitureRental()).toBeTruthy();
+  });
+});
