+++ front/src/app/core/models/furniture.spec.ts
@@ -1,0 +1,7 @@
+import { Furniture } from './furniture';
+
+describe('Furniture', () => {
+  it('should create an instance', () => {
+    expect(new Furniture()).toBeTruthy();
+  });
+});
