--- front/src/app/modules/models/furniture-rental/furniture-rental.component.html
@@ -1,1 +1,0 @@
-<p>furniture-rental works!</p>
