--- front/src/app/modules/models/furniture/furniture.component.html
@@ -1,1 +1,0 @@
-<p>furniture works!</p>
